Add a circle layer to the map

Add a circle layer to the map

add_circle_layer(map, source = NULL, filter = NULL,
  circle_blur = NULL, circle_color = NULL, circle_opacity = NULL,
  circle_pitch_alignment = NULL, circle_pitch_scale = NULL,
  circle_radius = NULL, circle_sort_key = NULL,
  circle_stroke_color = NULL, circle_stroke_opacity = NULL,
  circle_stroke_width = NULL, circle_translate = NULL,
  circle_translate_anchor = NULL, visibility = TRUE, popup = NULL,
  id = "circle-layer")

Arguments

map
A mapboxer object.

source
A Mapbox source. Uses the source from the mapboxer object if no source is supplied.

filter
A filter expression that is applied to the source.

circle_blur
(paint) Amount to blur the circle. 1 blurs the circle such that only the centerpoint is full opacity.

circle_color
(paint) The fill color of the circle.

circle_opacity
(paint) The opacity at which the circle will be drawn.

circle_pitch_alignment
(paint) Orientation of circle when map is pitched. One of "map", "viewport".

circle_pitch_scale
(paint) Controls the scaling behavior of the circle when the map is pitched. One of "map", "viewport".

circle_radius
(paint) The radius of the circle.

circle_sort_key
(layout) Sorts features in ascending order based on this value. Features with a higher sort key will appear above features with a lower sort key.

circle_stroke_color
(paint) The stroke color of the circle.

circle_stroke_opacity
(paint) The opacity of the circle's stroke.

circle_stroke_width
(paint) The width of the circle's stroke. Strokes are placed outside of the circle_radius.

circle_translate
(paint) The geometry's offset. Values are [x, y] where negatives indicate left and up, respectively.

circle_translate_anchor
(paint) Controls the frame of reference for circle_translate. One of "map", "viewport".

visibility
(layout) Whether the layer should be displayed.

popup
A mustache template in which the tags refer to the properties of the layer's data object.

id
The unique id of the layer.

See also

Examples

map <- quakes %>%
  as_mapbox_source(lng = "long", lat = "lat") %>%
  mapboxer(center = c(176.9, -24.655), zoom = 4) %>%
  add_navigation_control() %>%
  add_circle_layer(
    circle_color = "blue",
    circle_blur = 1,
    circle_stroke_color = "red",
    circle_stroke_width = 1,
    popup = "{{mag}}"
  ) %>%
  add_filter_control(
    "circle-layer",
    filter = list(">", "mag", 5),
    pos = "bottom-left"
  )

if (interactive()) map